requirements.txt
aiogram==3.0.0
gspread==5.7.2
google-auth==2.10.0
python-dotenv==1.0.0
